# docs/swagger.yaml
basePath: /api
definitions:
...
  structs.DeleteKategoriObatResponse:
    properties:
      pesan:
        example: Kategori obat berhasil dihapus
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.DispensingObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Dispensing obat berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.ErrorResponse:
    properties:
      pesan:
        example: Terjadi kesalahan
        type: string
      status:
        example: error
        type: string
    type: object
...
  structs.GetKategoriObatByIDResponse:
    properties:
      data:
        $ref: '#/definitions/structs.KategoriObat'
      pesan:
        example: Get kategori obat by ID berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.GetKategoriObatResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.KategoriObat'
        type: array
      pesan:
        example: Get kategori obat berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.GetTransaksiObatResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.TransaksiObat'
        type: array
      pesan:
        example: Get transaksi obat berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.KategoriObat:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deskripsi:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      modified_by:
        type: string
      nama_kategori:
        type: string
    type: object
  structs.LaporanTransaksiObat:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      kembalian:
        type: number
      nama_pasien:
        type: string
      no_resep:
        type: string
      status:
        type: string
      tipe_pembayaran:
        type: string
      total_harga:
        type: number
      total_item:
        type: integer
      total_pembayaran:
        type: number
    type: object
...
  structs.TambahKategoriObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.KategoriObat'
      pesan:
        example: Kategori obat berhasil ditambahkan
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.TambahdanUpdateKategoriObatRequest:
    properties:
      deskripsi:
        type: string
      nama_kategori:
        type: string
    required:
    - nama_kategori
    type: object
  structs.TransaksiObat:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      dibayar_at:
        type: string
      grand_total:
        type: number
      id:
        type: integer
      kembalian:
        type: number
      modified_at:
        type: string
      modified_by:
        type: string
      nama_pasien:
        type: string
      no_resep:
        type: string
      pasien_id:
        type: integer
      status:
        type: string
      tipe_pembayaran:
        type: string
      total_pembayaran:
        type: number
    type: object
...
  structs.UpdateKategoriObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.KategoriObat'
      pesan:
        example: Kategori obat berhasil diperbarui
        type: string
      status:
        example: sukses
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: '-'
    name: Damar
  description: 'Dokumentasi REST API Alur Obat Farmasi menggunakan Golang (Framework
    : Gin dan Database : PostgreSQL).'
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
  /api/admin-farmasi/kategori-obat:
    get:
      description: Mengambil Semua data kategori obat
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data kategori obat
          schema:
            $ref: '#/definitions/structs.GetKategoriObatResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Semua Kategori Obat
      tags:
      - 3. Kategori Obat
    post:
      consumes:
      - application/json
      description: Menambahkan kategori obat baru
      parameters:
      - description: Data Kategori Obat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahdanUpdateKategoriObatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Kategori obat berhasil ditambahkan
          schema:
            $ref: '#/definitions/structs.TambahKategoriObatResponse'
        "400":
          description: Gagal tambah kategori obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "409":
          description: Kategori obat sudah ada
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah Kategori Obat
      tags:
      - 3. Kategori Obat
  /api/admin-farmasi/kategori-obat/{id}:
    delete:
      description: Menghapus kategori obat
      parameters:
      - description: ID Kategori Obat
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Kategori obat berhasil dihapus
          schema:
            $ref: '#/definitions/structs.DeleteKategoriObatResponse'
        "400":
          description: ID Kategori obat tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Kategori obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Kategori Obat
      tags:
      - 3. Kategori Obat
    get:
      description: Mengambil detail kategori obat berdasarkan ID
      parameters:
      - description: ID Kategori Obat
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil detail kategori obat
          schema:
            $ref: '#/definitions/structs.GetKategoriObatByIDResponse'
        "400":
          description: ID Kategori Obat tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Kategori obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Kategori Obat By ID
      tags:
      - 3. Kategori Obat
    put:
      consumes:
      - application/json
      description: Mengubah data kategori obat
      parameters:
      - description: ID Kategori Obat
        in: path
        name: id
        required: true
        type: integer
      - description: Data Kategori Obat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahdanUpdateKategoriObatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Kategori obat berhasil diUpdate
          schema:
            $ref: '#/definitions/structs.UpdateKategoriObatResponse'
        "400":
          description: Gagal Update kategori obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Kategori obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Kategori Obat
      tags:
      - 3. Kategori Obat
...
  /api/apoteker/transaksi-obat/{id}/dispensing:
    put:
      description: Mengubah status transaksi obat dari PENDING menjadi DISPENSED.
        Transaksi harus memiliki minimal satu detail obat.
      parameters:
      - description: ID Transaksi Obat
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Proses dispensing transaksi resep obat (pemberian obat) berhasil
          schema:
            $ref: '#/definitions/structs.DispensingObatResponse'
        "400":
          description: ID Transaksi obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Gagal dispense transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dispensing Obat
      tags:
      - 8. Dispensed
...
  /api/apoteker/transaksi-obat/pending:
    get:
      description: Menampilkan daftar transaksi obat dengan status PENDING yang belum
        didispensing.
      produces:
      - application/json
      responses:
        "200":
          description: Get daftar pending transaksi resep obat berhasil
          schema:
            $ref: '#/definitions/structs.GetTransaksiObatResponse'
        "401":
          description: Data pending transaksi resep obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Transaksi Obat Pending
      tags:
      - 8. Dispensed
  /api/kepala-farmasi/laporan/transaksi-obat:
    get:
      description: Menampilkan laporan transaksi obat berdasarkan rentang tanggal.
      parameters:
      - description: Tanggal awal (YYYY-MM-DD)
        example: "2026-07-01"
        in: query
        name: tanggal_awal
        required: true
        type: string
      - description: Tanggal akhir (YYYY-MM-DD)
        example: "2026-07-31"
        in: query
        name: tanggal_akhir
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get daftar laporan transaksi resep obat pasien berhasil
          schema:
            $ref: '#/definitions/structs.LaporanTransaksiObat'
        "400":
          description: Transaksi obat tidak ditemukan pada rentang tanggal awal -
            akhir.
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Transaksi obat tidak ada.
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Laporan Transaksi Resep Obat
      tags:
      - 10. Laporan
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...